+++ src/main/kotlin/io/runescript/plugin/ide/config/RsConfig.kt
@@ -62,6 +62,7 @@ object RsConfig {
         "varphook",
         "stathook",
         "invhook",
+        "varchook",
         "any"
     )